finalização da construção do banco de dados, inicio da lógica de inserção e criação dos dados

+++ app/src/main/java/com/gabrielportari/workoutplannerapp/view/fragment/ManageWorkoutFragment.kt
@@ -1,5 +1,6 @@
 package com.gabrielportari.workoutplannerapp.view.fragment
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -46,7 +47,17 @@ class ManageWorkoutFragment : Fragment() {
             }
 
             override fun onDeleteClick(id: Int) {
-                viewModel.deleteWorkout(id)
+                val dialogBuilder: AlertDialog.Builder = AlertDialog.Builder(context)
+                dialogBuilder.setTitle("Deletar Treino")
+                dialogBuilder.setMessage("Tem certeza que deseja excluir esse treino?")
+                dialogBuilder.setPositiveButton("Sim") { _, _ ->
+                    viewModel.deleteWorkout(id)
+                }
+                dialogBuilder.setNegativeButton("Nao") { _, _ ->
+
+                }
+                val dialog = dialogBuilder.create()
+                dialog.show()
             }
 
             override fun onEditClick(workout: Workout) {
